HandelVisiter.php UI Overhaul
Due to inital design flaw (or lack of design) for this web page, a UI overhaul has been implemented. This overhaul only affects the UI and leaves the webpage's inputs and outputs unchanged.  The webpage is now designed to allow an unregistered person to complete registration by selecting a ticket type option, or allow a registered person to upgrade their ticket by selecting valid upgrade options.

In both cases, any ticket type options will be disabled if the ticket type is sold out. In addition priceing and remaining amount of tickets are shown for each ticket type option.